+++ FinalProject/FinalProject/DAL/MigratingCopyPasta.txt
@@ -1,3 +1,3 @@
-﻿
+﻿Enable-migrations -ContextTypeName JobPostingCFEntities -MigrationsDirectory DAL\Migrations
 add-migration -ConfigurationTypeName FinalProject.DAL.Migrations.Configuration
 udpate-database -ConfigurationTypeName FinalProject.DAL.Migrations.Configuration
