+++ FinalProject/FinalProject/Web.config
@@ -11,6 +11,8 @@
   <connectionStrings>
     <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-FinalProject-20180122100536.mdf;Initial Catalog=aspnet-FinalProject-20180122100536;Integrated Security=True"
       providerName="System.Data.SqlClient" />
+    <add name="JobPostingCFEntities" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-FinalProject-20180122100536.mdf;Initial Catalog=aspnet-FinalProject-20180122100536;Integrated Security=True"
+      providerName="System.Data.SqlClient" />
   </connectionStrings>
   <appSettings>
     <add key="webpages:Version" value="3.0.0.0" />
